pyproject.toml
[project]
name = "fastapi_auth_1"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Bllagden", email = "bllagden99@gmail.com"},
]
dependencies = [
    "fastapi[uvicorn]>=0.110.0",
    "uvicorn>=0.28.0",
    "sqlalchemy>=2.0.28",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "pydantic[email]==2.6.1",
    "pydantic-settings==2.1.0",
    "fastapi-cache2[redis]>=0.2.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.3.2",
    "black>=24.2.0",
    "isort>=5.13.2",
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.5.post1",
    "pytest-dotenv>=0.5.2",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
pythonpath = "src"
asyncio_mode = "auto"
env_files = [
    ".env.test",
]

[tool.isort]
profile = "black"

[tool.ruff]
fixable = [
  "F",
  "E",
  "W",
  "I",
  "COM",
  "UP",
  "RUF",
]
ignore = [
  "E501", # Line Length
  "A003", # Class attribute ... is shadowing a python builtin
  "ANN101",
  "ANN102",
  "D", # pydocstyle
  "INP",
  "TCH001",
  "TCH002",
  "TCH003",
  "RUF001",
  "RUF002",
  "RUF003",
  "RUF012",
  "RUF200",
  "ERA001",
  "TID252",
  "PLR0913",
]
select = ["ALL"]
src = ["src", "tests"]

[tool.ruff.per-file-ignores]
"src/alembic/*" = ["UP007"]
"src/logger.py" = ["ANN001", "UP008"]
"tests/*" = [
  "S101",
  "PT006", # Wrong name(s) type in `@pytest.mark.parametrize`, expected `tuple`
  "S311",
]

[tool.pdm.scripts]
lint_src = {composite = [
  "ruff check ./src --fix --show-fixes",
  "black ./src",
  "ruff check ./src --fix --show-fixes",
  "alembic check",
]}
